+++ lizard_auth_client/client.py
@@ -713,7 +713,7 @@ def get_billable_organisation(user, role_code=models.Role.BILLING_ROLE_CODE):
                 synced = True
 
 
-def sso_server_url(name, ignore_cache=False):
+def sso_server_url(name, use_cache=True):
     """Return url of endpoint on the SSO server
 
     The v2 API has a starting point that lists the available endpoints. We
@@ -733,7 +733,7 @@ def sso_server_url(name, ignore_cache=False):
     cache_key = 'cached_sso_server_urls'
     sso_server_urls = cache.get(cache_key)
 
-    if sso_server_urls is None or ignore_cache:
+    if sso_server_urls is None or not use_cache:
         # import here so this module can easily be reused outside of Django
         from lizard_auth_client.conf import settings
         # First time, grab it from the server.